Improving script readability. Load/unload the model per case (not per repeat).
